chat-service/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
boto3
websockets
